Add some methods to JdbcCardsService

+++ hackathon/src/main/java/com/greencats/dto/card/CardEditInfo.java
@@ -1,7 +1,7 @@
 package com.greencats.dto.card;
 
 public record CardEditInfo (
-    Integer card_id,
+    Long card_id,
     Integer complexity
 ) {
